refactoring endpoint classes
