--- lib/src/main/java/hajimeapi4j/internal/datatype/utilizations/Arrange.java
@@ -1,18 +1,0 @@
-package hajimeapi4j.internal.datatype.utilizations;
-
-import hajimeapi4j.api.endpoint.EndPoint;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.ToString;
-
-/**
- *
- */
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-@EqualsAndHashCode
-@ToString
-public class Arrange implements EndPoint {
-
-
-}
