--- lib/src/main/java/hajimeapi4j/internal/datatype/utilizations/Composer.java
@@ -1,5 +1,0 @@
-package hajimeapi4j.internal.datatype.utilizations;
-
-public class Composer {
-
-}
